CR - Move CSS inside of component folder
