+++ apps/site/src/app/_components/NotificationDialog/NotificationDialog.css
@@ -1,0 +1,27 @@
+@keyframes slide-in-from-top {
+    from {
+      transform: translateY(-100%);
+    }
+    to {
+      transform: translateY(0);
+    }
+  }
+  
+  @keyframes shrink-and-fade-out {
+    from {
+      transform: scale(1);
+      opacity: 1;
+    }
+    to {
+      transform: scale(0.9);
+      opacity: 0;
+    }
+  }
+  
+  .animate-slide-in-from-top {
+    animation: slide-in-from-top 300ms cubic-bezier(0.41, 0.73, 0.51, 1.02);
+  }
+  
+  .animate-shrink-and-fade-out {
+    animation: shrink-and-fade-out 150ms ease-in forwards;
+  }
