+++ apps/site/src/app/_styles/globals.css
@@ -21,30 +21,6 @@
   --color-brand-800: #08072e;
 
   --spacing-readable: 60ch;
-
-  --animate-slide-in-from-top: slide-in-from-top 300ms
-    cubic-bezier(0.41, 0.73, 0.51, 1.02);
-  --animate-shrink-and-fade-out: shrink-and-fade-out 150ms ease-in forwards;
-
-  @keyframes slide-in-from-top {
-    from {
-      transform: translateY(-100%);
-    }
-    to {
-      transform: translateY(0);
-    }
-  }
-
-  @keyframes shrink-and-fade-out {
-    from {
-      transform: scale(1);
-      opacity: 1;
-    }
-    to {
-      transform: scale(0.9);
-      opacity: 0;
-    }
-  }
 }
 
 @utility brand-outline {
